+++ .gitignore
@@ -25,3 +25,6 @@ node_modules
 
 # build file
 dist
+
+# environment variables
+.env
