+++ src/App.jsx
@@ -1,7 +1,6 @@
 import Home from "./pages/Home";
 import { Route, Routes } from "react-router-dom";
 import About from "./pages/About";
-import SuccessPage from "./pages/SuccessPage";
 import Navbar from "./components/Navbar";
 import Footer from "./components/Footer";
 import { useLocation } from "react-router-dom";
@@ -15,7 +14,6 @@ function App() {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/about" element={<About />} />
-        <Route path="/success" element={<SuccessPage />} />
       </Routes>
       <Footer />
     </div>
